# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de denuncias",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/complaints": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "All complaints",
                "description": "All complaints",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": ""
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "departments": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "anonymous": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "region": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "province": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "codmunicipalitye": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "file": {
                                            "type": "file",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/complaints/person/{identification}": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "All complaints of person",
                "description": "All complaints of person",
                "operationId": "complaints_person",
                "parameters": [
                    {
                        "name": "identification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "identification": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": ""
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "departments": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "anonymous": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "region": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "province": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "codmunicipalitye": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "file": {
                                            "type": "file",
                                            "example": ""
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/complaint/delete/{id}": {
            "delete": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Delete complaints",
                "description": "Delete complaints",
                "operationId": "complaint_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}